attn_res: keep the vendored read file free of our own code
The SPDX tag governs a whole file, so the shims sharing one with the vendored
function put Tenstorrent code under the Kimi K3 License. Moving them to
hf_walk.py makes the license boundary the file boundary.
